requirement.txt
psutil
Pillow

